+++ Domain.Empresas/Repositories/IEnterpriseRepository.cs
@@ -12,16 +12,16 @@ namespace Domain.Empresas.Repositories
         Enterprise LoadNavigation(Enterprise enterprise, Expression<Func<Enterprise, object>> expression);
 
         Enterprise LoadCollection(Enterprise enterprise, string collectionPropName);
-        Enterprise GetWithKeys(long id);
+        Enterprise GetWithKeys(int id);
 
-        Enterprise GetWithKeys(long id, IEnumerable<string> navigations = null, IEnumerable<string> collections = null);
+        Enterprise GetWithKeys(int id, IEnumerable<string> navigations = null, IEnumerable<string> collections = null);
         IQueryable<Enterprise> GetAll(bool readOnly = false);
 
         IQueryable<Enterprise> GetAll(string included = "", bool readOnly = false);
         Enterprise Insert(Enterprise entity);
         Enterprise Remove(Enterprise entity);
 
-        Enterprise Remove(long id);
+        Enterprise Remove(int id);
         IQueryable<Enterprise> Query(Expression<Func<Enterprise, bool>> predicate = null, bool readOnly = false, string included = "");
     }
 }
